Simplify handling of super sends and fix issues
This removes the use of Single, which was a box for a flag. Instead, I am now using return values and passing on the flag.

The underlying problem was that more than the first message could be marked as super send, which is wrong, I believe.
For binary messages, only the second argument would show this behavior, so it was asymmetric.

Signed-off-by: Stefan Marr <git@stefan-marr.de>